Replace simple token authenication gem with an encrypted API token field.
It's bothered me for a while that the tokens were plaintext in the
database. This change will reset everyone's tokens, but that
shouldn't be a huge problem since there aren't many people using the API
right now anyway.

I still need to get all the tests working and probably write a few more.
For some reason the app doesn't accept the token in the current spec
I've written.