Provide the necessary headers to GraphiQL by default.
This isn't _great_, but it'll do. The problem is that there's no way to
get the docs explorer to reload when the headers change, so they have to
be available on first load. The docs explorer will fail for users that
aren't logged in, but I can live with that.

And now GraphiQL will actually render in production!