Use variants for images in the GraphQL API.
This is a breaking change because the GraphQL API now returns
the full URL rather than just the path, but I'm pretty sure
only one person other than myself is using the endpoint anyway.

This change makes the images returned by the API much smaller by
default.

This will also allow the URL to be swapped out for a CDN in the future
without breaking the API for end users.