Update Sorbet and exclude faker from tapioca.
This way, the autogenerated Faker RBI in sorbet-typed will be used by
Sorbet and show the method documentation on-hover.

I was hoping the fix in the below PR would resolve the issue of
documentation not showing up for the methods, but for whatever reason
Sorbet's file ordering doesn't seem to go in alphabetical order? So this
seems to be the only way to fix the problem.

https://github.com/sorbet/sorbet/pull/3895