Add fallbacks to the secret key and salt, only used in development.
In production (and test), the fallbacks are set to nil and will raise an
error if hit. This is good so that we can never run the application in
production if it's got no secret key set.

I noticed this after trying to regenerate Sorbet types when the master
key wasn't available.