Update the GraphQL types to fix incorrect nullability.
There are a bunch of places where I set an attribute in GraphQL as
nullable when I really shouldn't have, because it'll always return an
array (although that array may be empty).

This was causing annoying typechecking issues in the frontend app, so I
think it's worth making a "breaking" change here. For all of these,
there's no change to the underlying functionality, they were never
returning `null` anyway. The only type where that wasn't the case is
User, and so User is the only one that really has a breaking change.