Upgrade `load_defaults` config to Rails 6.1.
Apparently it doesn't seem to like that I still had the
`new_framework_defaults_6_1.rb` file, and now that I've removed that it
shuts up about some of the deprecation warnings. I verified that I
upgraded to all the 6.1 defaults in that file, so deleting it and
telling Rails to use the 6.1 defaults should be fine.