Revert graphiql package upgrade.
v2.3.0 of graphiql breaks my integration with GraphiQL for some reason.
We don't want that, so just downgrade it for now.

I think it's https://github.com/graphql/graphiql/pull/2939 that broke
this, but I'm not 100% sure. I know I'm loading graphiql in a weird way,
so that is probably part of the problem.