Disable game creation and updating for normal users.
Moderators and admins are stil able to create and edit games, but
not normal users. I'm making this change to prevent any chance of
vandalism, since >99% of the edits on the site are automated via
Wikidata imports anyway.

If an attacker chose to vandalize the site by editing dozens or hundreds
of game records, they could do serious damage that would be difficult to
reverse (most likely I'd need to restore the games table from a database
backup). This would take a good amount of time to fix and potentially
result in the site either going down temporarily or even in data loss of
recent, legitimate changes due to the database being reverted to an
older state.

Since the edit rate by users is pretty low, this shouldn't be a
significant disruption to site usage, and I've chosen not to implement
an approval system since that'd require a lot of new code and
maintenance work long-term.

Editing games can still be done by anyone, by creating/editing the
corresponding records on Wikidata. They'll be picked up automatically
the next time the Wikidata import script is run.